Focus on scripting system. WIP

+++ editor/scripting/robot2D-ScriptCore/robot2D-ScriptCore/Scene/Collision2D.cs
@@ -10,9 +10,8 @@ namespace robot2D
     
     public class Collision2D
     {
-        internal uint m_CameraCollider;
-        
-        internal uint m_otherCollider;
+        internal ulong m_CameraCollider;
+        internal ulong m_otherCollider;
 
         public Collider2D collider
         {
